:art: Refactor: Correct package naming from 'modele' to 'model' for clothing and tag classes
