+++ src/main/java/com/example/project/repository/clothingLib/TagRepository.java
@@ -1,7 +1,8 @@
 package com.example.project.repository.clothingLib;
 
-import com.example.project.modele.clothingLib.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.example.project.model.clothingLib.Tag;
 
 public interface TagRepository extends JpaRepository<Tag, Long> {
 }
